requirements.txt
flask==2.0.1
pymongo==3.12.0
flask-login==0.5.0
passlib==1.7.4
python-dotenv==0.19.0
requests==2.26.0
flask-wtf==0.15.1
wtforms==2.3.3